requirements.txt
# Fal Gram Bot - Requirements for Render.com Deployment
# Version: 3.1.1

# Core dependencies
python-telegram-bot==21
python-dotenv==1.0.0
supabase==2.5
google-generativeai==0.8.3

# Web service dependencies
flask==3.0.0
gunicorn==21.2.0

# Async and scheduling
apscheduler==3.10.4
asyncio==3.4.3

# PDF generation
fpdf2==2.7.8

# HTTP client
httpx==0.27.0
requests==2.31.0

# Database and utilities
psycopg2-binary==2.9.9
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
structlog==23.2.0

# Security
cryptography==41.0.7


